Readjusting method file
